+++ cli/lib/command/internal_command_pull.dart
@@ -9,10 +9,9 @@ import 'package:stax/external_command/extended_process_result.dart';
 import 'package:stax/git/branch_info.dart';
 
 class InternalCommandPull extends InternalCommand {
-  static final stayOnMainFlag = Flag(
-    short: '-m',
-    long: '--stay-on-main',
-    description: 'Stay on the main/target branch after pulling.',
+  static final stayOnHeadFlag = Flag(
+    long: '--stay-on-head',
+    description: 'Stay on the head/default branch after pulling.',
   );
 
   InternalCommandPull()
@@ -25,7 +24,7 @@ class InternalCommandPull extends InternalCommand {
         },
         flags: [
           InternalCommandDeleteStale.forceDeleteFlag,
-          stayOnMainFlag,
+          stayOnHeadFlag,
           InternalCommandDeleteStale.skipDeleteFlag,
         ],
       );
@@ -40,7 +39,7 @@ class InternalCommandPull extends InternalCommand {
     );
     final hasForceDeleteFlag = InternalCommandDeleteStale.forceDeleteFlag
         .hasFlag(args);
-    final hasStayOnMainFlag = stayOnMainFlag.hasFlag(args);
+    final hasStayOnHeadFlag = stayOnHeadFlag.hasFlag(args);
     final currentBranch = context.getCurrentBranch();
     final targetBranch = args.elementAtOrNull(0);
     final defaultBranch = targetBranch ?? context.getDefaultBranch();
@@ -72,7 +71,7 @@ class InternalCommandPull extends InternalCommand {
             .printNotEmptyResultFields()
             .assertSuccessfulExitCode();
     if (result == null) {
-      if (!hasStayOnMainFlag && needToSwitchBranches && currentBranch != null) {
+      if (!hasStayOnHeadFlag && needToSwitchBranches && currentBranch != null) {
         (await context.git.switch0
                 .arg(currentBranch)
                 .announce("Switching back to original branch '$currentBranch'.")
@@ -126,7 +125,7 @@ class InternalCommandPull extends InternalCommand {
       }
     }
 
-    if (!hasStayOnMainFlag &&
+    if (!hasStayOnHeadFlag &&
         (needToSwitchBranches || additionalBranches.isNotEmpty) &&
         currentBranch != null) {
       (await context.git.switch0