+++ cli/test/cli/help_test.dart
@@ -74,8 +74,8 @@ Note: you can type first letter or couple of first letters instead of full comma
          opt1 - Optional target branch, will default to <remote>/HEAD
       Flags:
          -f, --force-delete - Force delete gone branches.
-         -m, --stay-on-main - Stay on the main/target branch after pulling.
          -s, --skip-delete - Skip deletion of gone branches.
+         --stay-on-head - Stay on the head/default branch after pulling.
  • pull-request, pr - Creates a pull request.
  • rebase - rebase tree of branches on top of main
       Positional arguments: